+++ server_2x/package.json
@@ -27,10 +27,12 @@
     "@nestjs/jwt": "^7.2.0",
     "@nestjs/passport": "^7.1.5",
     "@nestjs/platform-express": "^7.5.1",
+    "@nestjs/platform-ws": "^7.6.12",
     "@nestjs/serve-static": "^2.1.4",
     "@nestjs/swagger": "^4.7.13",
     "@nestjs/terminus": "^7.1.0",
     "@nestjs/typeorm": "^7.1.5",
+    "@nestjs/websockets": "^7.6.12",
     "bcrypt": "^5.0.0",
     "class-transformer": "^0.4.0",
     "class-validator": "^0.13.1",
@@ -61,6 +63,7 @@
     "@types/passport-jwt": "^3.0.4",
     "@types/passport-local": "^1.0.33",
     "@types/supertest": "^2.0.10",
+    "@types/ws": "^7.4.0",
     "@typescript-eslint/eslint-plugin": "^4.6.1",
     "@typescript-eslint/parser": "^4.6.1",
     "eslint": "^7.12.1",
