+++ server_2x/src/octoprint/octoprint.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { OctoprintService } from './services/octoprint.service';
+import { OctoprintGateway } from './gateway/octoprint.gateway';
+
+@Module({
+  providers: [OctoprintGateway, OctoprintService]
+})
+export class OctoprintModule {}
