+++ server_2x/src/octoprint/services/octoprint.service.ts
@@ -1,0 +1,5 @@
+import {Injectable} from '@nestjs/common';
+
+@Injectable()
+export class OctoprintService {
+}
